+++ packages/dara-components/changelog.md
@@ -5,6 +5,8 @@ title: Changelog
 ## 1.18.10
 
 -  Fixed an issue where `Table` type inference would not handle the different datetime64 types correctly.
+-  Fixed an issue where `SectionedList` would not update the input value when clicking on an item.
+-  Updated the `SectionedList` component to support placeholder text.
 
 ## 1.16.23